src/PrayerWS/requirements.txt
django
djangorestframework
djangorestframework-gis
geocoder
pyproj
requests
openpyxl
